Refactoring AgentRepository (it's part of #1368)
### What's done:
* extracted AgentRepository from AgentStatusRepository
* replaced all BodilessResponseEntity to EmptyResponse
* added additional method in AgentsController
* removed FsUtils from orchestrator
