+++ save-backend/src/main/kotlin/com/saveourtool/save/backend/repository/AgentStatusRepository.kt
@@ -1,10 +1,5 @@
-/**
- * JPA repositories for Agent related data
- */
-
 package com.saveourtool.save.backend.repository
 
-import com.saveourtool.save.entities.Agent
 import com.saveourtool.save.entities.AgentStatus
 import com.saveourtool.save.spring.repository.BaseEntityRepository
 import org.springframework.stereotype.Repository
@@ -36,42 +31,4 @@ interface AgentStatusRepository : BaseEntityRepository<AgentStatus> {
      */
     @Transactional
     fun deleteByAgentExecutionIdIn(ids: List<Long>)
-}
-
-/**
- * JPA repository for agents.
- */
-@Repository
-interface AgentRepository : BaseEntityRepository<Agent> {
-    /**
-     * Find agent by its agent id
-     *
-     * @param agentId agent id
-     * @return [Agent]
-     */
-    fun findByContainerId(agentId: String): Agent?
-
-    /**
-     * Find all agents with [executionId]
-     *
-     * @param executionId id of execution
-     * @return list of agents
-     */
-    fun findByExecutionId(executionId: Long): List<Agent>
-
-    /**
-     * Find all agents with [projectId] in execution
-     *
-     * @param projectId id of project
-     * @return list of agents
-     */
-    fun findByExecutionProjectId(projectId: Long): List<Agent>
-
-    /**
-     * Delete all agents with [executionId]
-     *
-     * @param ids list id of execution
-     */
-    @Transactional
-    fun deleteByExecutionIdIn(ids: List<Long>)
 }
