--- save-orchestrator-common/src/main/kotlin/com/saveourtool/save/orchestrator/utils/FsUtils.kt
@@ -1,44 +1,0 @@
-/**
- * Methods to work with file system
- */
-
-package com.saveourtool.save.orchestrator.utils
-
-import com.saveourtool.save.utils.warn
-import java.nio.file.Files
-import java.nio.file.LinkOption
-import java.nio.file.Path
-import java.nio.file.attribute.PosixFileAttributeView
-import java.nio.file.attribute.PosixFilePermission
-import kotlin.io.path.name
-
-internal val allExecute = setOf(PosixFilePermission.OWNER_EXECUTE, PosixFilePermission.GROUP_EXECUTE, PosixFilePermission.OTHERS_EXECUTE)
-
-/**
- * Try to add executable permissions for file denoted by [this]
- */
-context(LoggingContext)
-internal fun Path.tryMarkAsExecutable() {
-    try {
-        Files.setPosixFilePermissions(this, Files.getPosixFilePermissions(this) + allExecute)
-    } catch (ex: UnsupportedOperationException) {
-        logger.warn(ex) { "Failed to mark file ${this.name} as executable" }
-    }
-}
-
-/**
- * Change owner of all files under [this] to user named [user]
- *
- * @receiver target directory
- * @param user name of the new owner
- */
-@Suppress("KDOC_NEWLINES_BEFORE_BASIC_TAGS")  // will be fixed in diktat 1.2.2
-internal fun Path.changeOwnerRecursively(user: String) {
-    val lookupService = fileSystem.userPrincipalLookupService
-    toFile().walk().forEach { file ->
-        Files.getFileAttributeView(file.toPath(), PosixFileAttributeView::class.java, LinkOption.NOFOLLOW_LINKS).apply {
-            setGroup(lookupService.lookupPrincipalByGroupName(user))
-            setOwner(lookupService.lookupPrincipalByName(user))
-        }
-    }
-}
